+++ ios/MullvadVPN/UI appearance/UIColor+Palette.swift
@@ -36,6 +36,17 @@ extension UIColor {
         static let invalidInputTextColor = UIColor.dangerColor
     }
 
+    enum SearchTextField {
+        static let placeholderTextColor = TextField.placeholderTextColor
+        static let inactivePlaceholderTextColor = UIColor.white
+        static let textColor = TextField.textColor
+        static let inactiveTextColor = UIColor.white
+        static let backgroundColor = TextField.backgroundColor
+        static let inactiveBackgroundColor = UIColor.secondaryColor
+        static let leftViewTintColor = UIColor.primaryColor
+        static let inactiveLeftViewTintColor = UIColor.white
+    }
+
     enum AppButton {
         static let normalTitleColor = UIColor.white
         static let highlightedTitleColor = UIColor.lightGray